feat(mobile): redesign navigation bar with tooltips and shared News component
- Restyle mobile nav bar with bg image, uniform button sizing, hover states, and tooltips
- Portal Tooltip content to prevent layout shift on hover
- Extract shared useNews hook and NewsTooltipBody so desktop and mobile reuse the same News dialog, indicator dot, and unread-posts tooltip
- Switch News button layout to flex-col with responsive label ("News" mobile, "News & Updates" desktop)
- Drop standalone mobile/news component in favor of the shared one
