+++ src/containers/navigation/mobile/index.tsx
@@ -2,19 +2,21 @@ import ConfigureWidgets from '@/containers/navigation/mobile/configure';
 import LanguageSelector from '@/containers/navigation/mobile/language-selector';
 import Locations from '@/containers/navigation/mobile/locations';
 import MapToggle from '@/containers/navigation/mobile/map-toggle';
-import News from '@/containers/navigation/mobile/news';
+import News from '@/containers/navigation/news';
+
+import { TooltipProvider } from '@/components/ui/tooltip';
 
 const NavigationMobile = () => {
   return (
-    <div className="pointer-events-auto fixed -bottom-0.5 z-50 h-36 w-screen bg-[url(/images/mobile-sidebar-bg.svg)] bg-cover bg-no-repeat">
-      <div className="mt-20 flex h-24 max-w-screen items-center justify-center space-x-6">
+    <TooltipProvider delayDuration={200}>
+      <div className="bg-brand-600 pointer-events-auto fixed bottom-0 left-0 z-50 flex h-20 w-screen items-center justify-center space-x-6 bg-[url(/images/mobile-sidebar-bg.svg)] bg-cover bg-bottom bg-no-repeat">
         <Locations />
         <News />
         <ConfigureWidgets />
         <LanguageSelector />
         <MapToggle />
       </div>
-    </div>
+    </TooltipProvider>
   );
 };
