+++ src/containers/navigation/mobile/locations/index.tsx
@@ -9,6 +9,7 @@ import {
   DialogClose,
   DialogTitle,
 } from '@/components/ui/dialog';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
 
 import GLASS_SVG from '@/svgs/sidebar/glass';
 
@@ -24,32 +25,37 @@ const LocationsMobile = () => {
   }, []);
 
   return (
-    <div className="flex h-full flex-col space-y-1.5 text-center">
-      <Dialog open={isOpen}>
-        <DialogTrigger asChild>
-          <button
-            onClick={openMenu}
-            className="mt-1 flex cursor-pointer items-center justify-center rounded-full"
-          >
-            <GLASS_SVG
-              className="h-8 w-8 fill-current fill-white stroke-white"
-              role="img"
-              title="Glass"
-            />
-          </button>
-        </DialogTrigger>
-        <DialogContent
-          className="top-0 min-h-screen w-screen rounded-none pt-4 pb-0"
-          onEscapeKeyDown={closeMenu}
-          onInteractOutside={closeMenu}
-        >
-          <DialogTitle className="sr-only">Locations</DialogTitle>
-          <LocationsList onSelectLocation={closeMenu} />
-          <DialogClose onClose={closeMenu} className="top-4" />
-        </DialogContent>
-      </Dialog>
-      <div className="text-xxs font-sans leading-[10px] text-white">Locations</div>
-    </div>
+    <Dialog open={isOpen}>
+      <Tooltip>
+        <TooltipTrigger asChild>
+          <DialogTrigger asChild>
+            <button
+              type="button"
+              onClick={openMenu}
+              aria-label="Locations"
+              className="flex w-14 cursor-pointer flex-col items-center justify-center space-y-1 text-white transition-opacity hover:opacity-80 focus-visible:opacity-80 focus-visible:outline-none"
+            >
+              <GLASS_SVG
+                className="h-8 w-8 fill-current fill-white stroke-white"
+                role="img"
+                title="Glass"
+              />
+              <span className="text-xxs font-sans leading-none text-white">Locations</span>
+            </button>
+          </DialogTrigger>
+        </TooltipTrigger>
+        <TooltipContent side="top">Search locations</TooltipContent>
+      </Tooltip>
+      <DialogContent
+        className="top-0 min-h-screen w-screen rounded-none pt-4 pb-0"
+        onEscapeKeyDown={closeMenu}
+        onInteractOutside={closeMenu}
+      >
+        <DialogTitle className="sr-only">Locations</DialogTitle>
+        <LocationsList onSelectLocation={closeMenu} />
+        <DialogClose onClose={closeMenu} className="top-4" />
+      </DialogContent>
+    </Dialog>
   );
 };
